Enhancement: Extract StringStartsWithTest
